Added Controller Input
At a base level controller input works, however it currentley overrides keyboard input and  when using DS4 Windows, will not allow the user to return to keyboard input once the controller is disconnected. To resolve, Unity must be closed and re-opened. Will track this issue and provide a fix later on.